server/requirements.txt
openenv-core[core]>=0.2.1
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=1.0.0
torch>=2.11.0
executorch>=1.2.0
numpy>=1.26.0
